rename batchnorm files to norm
