+++ tests/nn/test_norm.py
